Added Filter: "Environment" / Deleted Unused Files
The environment filter/file will contain all information pertaining to
the Environment.  Specifically an Environment will contain classes,
data-structures, and methods that codify the floor layout (walls, doors,
elevators, exits, stairs... etc) and the locations and types of sensors,
alarms, runway exit lights, etc...  Essentially it contains the "map" of
the building as well as the tools to load and save the "map".